src/models/terramind.py
import torch
import torch.nn as nn
from typing import Dict, Any, Optional
from .base_model import BaseGeospatialModel

class TerraMindModel(BaseGeospatialModel):
    """TerraMind model wrapper for geospatial tasks."""
    
    def __init__(self, config: Dict[str, Any]):
        super().__init__(config)
        self.model_name = 'TerraMind-1.0'
        self.hf_model_name = 'ibm-esa-geospatial/TerraMind-1.0-large'
        
        # TerraMind specific configuration
        self.modalities = config.get('modalities', ['S2L2A', 'S1GRD'])
        self.merge_method = config.get('merge_method', 'mean')
        
        # Initialize TerraMind backbone (placeholder for actual implementation)
        self._init_terramind_backbone()
        
    def _init_terramind_backbone(self):
        """Initialize TerraMind backbone model."""
        try:
            # This would use the actual TerraTorch integration
            from terratorch import BACKBONE_REGISTRY
            self.backbone = BACKBONE_REGISTRY.build(
                'terramind_v1_large',
                pretrained=True,
                modalities=self.modalities
            )
        except ImportError:
            # Fallback implementation for demonstration
            print("TerraTorch not available. Using placeholder implementation.")
            self.backbone = nn.Sequential(
                nn.Conv2d(12, 768, kernel_size=16, stride=16),  # Patch embedding
                nn.Flatten(2),
                nn.Transpose(1, 2)
            )
            
    def forward(self, x: torch.Tensor, **kwargs) -> torch.Tensor:
        """Forward pass through TerraMind model."""
        if isinstance(x, dict):
            # Multi-modal input
            return self.backbone(x)
        else:
            # Single modality input
            return self.backbone(x)
    
    def get_features(self, x: torch.Tensor, layer_names: Optional[list] = None) -> Dict[str, torch.Tensor]:
        """Extract intermediate features for knowledge distillation."""
        features = {}
        
        # TerraMind specific feature extraction
        output = self.forward(x)
        features['patch_embeddings'] = output
        
        # Additional feature extraction based on layer names
        if layer_names:
            # Implementation would depend on TerraMind architecture
            pass
            
        return features
    
    @classmethod
    def from_pretrained(cls, model_path: str = None, **kwargs):
        """Load pretrained TerraMind model."""
        config = kwargs.get('config', {})
        config['model_name'] = 'TerraMind-1.0'
        return cls(config)
    
    def generate(self, input_modalities: Dict[str, torch.Tensor], 
                output_modalities: list, timesteps: int = 10) -> Dict[str, torch.Tensor]:
        """Generate output modalities using TerraMind's any-to-any capability."""
        # This would use TerraMind's generation capabilities
        # Placeholder implementation
        generated = {}
        for modality in output_modalities:
            # Generate synthetic output for demonstration
            batch_size = list(input_modalities.values())[0].shape[0]
            generated[modality] = torch.randn(batch_size, 3, 224, 224)
        return generated