requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.21.0
scipy>=1.7.0

# Geospatial and Earth observation
rasterio>=1.3.0
geopandas>=0.13.0
xarray>=2023.1.0
netcdf4>=1.6.0
satellite-image-deep-learning>=0.1.0

# Machine learning and data science
scikit-learn>=1.3.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
pillow>=9.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0
click>=8.0.0
loguru>=0.7.0

# Optional: TerraTorch integration
# terratorch>=0.1.0  # Uncomment when available

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0